backend/pyproject.toml
[tool.poetry]
name = "backend"
version = "0.1.0"
description = ""
authors = ["Levenko Evgeny <levenko.evgeny@yandex.by>"]

[tool.poetry.dependencies]
python = "^3.8"
Django = "^4.0.5"
djangorestframework = "^3.13.1"
django-filter = "^22.1"
django-cors-headers = "^3.13.0"
django-user-agents = "^0.4.0"
djangorestframework-simplejwt = "^5.2.0"
python-jose = "^3.3.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
